+++ CHANGELOG.md
@@ -5,3 +5,8 @@ First release.
 ### 0.2.0
 
 * API change: change last 2 write method arguments to keyword arguments.
+
+### 0.3.0
+
+* Parser can now be configured with array name, record name, and numeric field list.
+* Interface has changed, must instantiate FideParser for FIDE files.